refactor ui inspector sections
