+++ crates/rulemorph_ui/ui/src/inspector_section_types.ts
@@ -1,0 +1,11 @@
+export type TraceInspectorSections = {
+  finalize: boolean;
+  step: boolean;
+  opList: boolean;
+  opResult: boolean;
+};
+
+export type ApiInspectorSections = {
+  opList: boolean;
+  memo: boolean;
+};
